refactor(#18): FSD 구조 이관 및 레거시 브릿지 제거
