+++ src/entities/notice/model/types.ts
@@ -37,22 +37,3 @@ export interface ParsedConditions {
 export interface ParsedNotice extends Notice {
   conditions: ParsedConditions;
 }
-
-export interface UserProfile {
-  age: number;
-  maritalStatus: "single" | "married" | "newlywed";
-  householdSize: number;
-  currentRegion: string;
-  noHomeYears: number;
-  income: number;
-  asset: number;
-  carAsset: number;
-  subscriptionDate: string;
-  subscriptionCount: number;
-  subscriptionAmount: number;
-  regions: string[];
-  minArea: number;
-  maxArea: number;
-  minBuildYear: number;
-  housingTypes: string[];
-}
