+++ src/features/collect-notices/index.ts
@@ -1,0 +1,1 @@
+export * from "./model/collect-notices";
