+++ src/features/env-setup/model/constants.ts
@@ -1,0 +1,36 @@
+export const REGION_CHOICES = [
+  { title: "전국", value: "00" },
+  { title: "서울", value: "11" },
+  { title: "경기", value: "41" },
+  { title: "인천", value: "28" },
+  { title: "부산", value: "26" },
+  { title: "대구", value: "27" },
+  { title: "광주", value: "29" },
+  { title: "대전", value: "30" },
+  { title: "울산", value: "31" },
+  { title: "세종", value: "36" },
+  { title: "강원", value: "42" },
+  { title: "충북", value: "43" },
+  { title: "충남", value: "44" },
+  { title: "전북", value: "45" },
+  { title: "전남", value: "46" },
+  { title: "경북", value: "47" },
+  { title: "경남", value: "48" },
+  { title: "제주", value: "50" },
+];
+
+export const HOUSING_TYPE_CHOICES = [
+  { title: "임대주택 (06)", value: "06" },
+  { title: "매입/전세임대 (13)", value: "13" },
+  { title: "공공분양 (05)", value: "05" },
+  { title: "토지/분양 (01)", value: "01" },
+  { title: "상가/업무시설 (22)", value: "22" },
+];
+
+export const AREA_PRESET_CHOICES = [
+  { title: "소형 39㎡ (약 16평형)", value: 39 },
+  { title: "소형 49㎡ (약 20평형)", value: 49 },
+  { title: "국민평형 59㎡ (약 24평형)", value: 59 },
+  { title: "중형 74㎡ (약 30평형)", value: 74 },
+  { title: "중형 84㎡ (약 34평형)", value: 84 },
+];
