+++ src/features/notify-slack/index.ts
@@ -1,0 +1,1 @@
+export * from "./model/send-slack-notification";
