+++ src/features/parse-notice/index.ts
@@ -1,0 +1,1 @@
+export * from "./model/parse-notices";
